Limit MACOSX_DEPLOYMENT_TARGET based on Python.
It's not possible to compile Python extensions targeting an earlier SDK than
Python itself was compiled for. Conversely, we don't want to target a later
version of the SDK than is available on the current platform. Finally, the
stack itself requires at least OSX 10.9.

Here, we attempt to be smart about choosing an SDK version that satisfies all
the above constraints.